refactor: split config and state from api
