+++ tests/test_connect.py
@@ -3,7 +3,7 @@ from fastapi import HTTPException
 
 
 @pytest.mark.asyncio
-async def test_connect_error(monkeypatch, bridge):
+async def test_connect_error(monkeypatch, state_module):
     class FailClient:
         def __init__(self, *args, **kwargs):
             self.host = kwargs["host"]
@@ -12,10 +12,10 @@ async def test_connect_error(monkeypatch, bridge):
         def connect(self, callback=None):
             raise RuntimeError("boom")
 
-    monkeypatch.setattr(bridge, "BambuClient", FailClient)
+    monkeypatch.setattr(state_module, "BambuClient", FailClient)
     with pytest.raises(HTTPException) as excinfo:
-        await bridge._connect("p1")
+        await state_module._connect("p1")
     assert excinfo.value.status_code == 502
-    _, errors = await bridge.state.snapshot()
+    _, errors = await state_module.state.snapshot()
     assert "p1" in errors
     assert "RuntimeError" in errors["p1"]
