+++ tests/test_env_helpers.py
@@ -1,19 +1,19 @@
 import pytest
 
 
-def test_pairs(monkeypatch, bridge):
+def test_pairs(monkeypatch, cfg):
     monkeypatch.setenv("TEST_PAIRS", "a@1;b@2")
-    assert bridge._pairs("TEST_PAIRS") == {"a": "1", "b": "2"}
+    assert cfg._pairs("TEST_PAIRS") == {"a": "1", "b": "2"}
 
 
-def test_kv(monkeypatch, bridge):
+def test_kv(monkeypatch, cfg):
     monkeypatch.setenv("TEST_KV", "a=1;b=2")
-    assert bridge._kv("TEST_KV") == {"a": "1", "b": "2"}
+    assert cfg._kv("TEST_KV") == {"a": "1", "b": "2"}
 
 
-def test_validate_env_missing(bridge, monkeypatch):
-    monkeypatch.setattr(bridge, "PRINTERS", {"p1": "h"})
-    monkeypatch.setattr(bridge, "SERIALS", {})
-    monkeypatch.setattr(bridge, "LAN_KEYS", {"p1": "k"})
+def test_validate_env_missing(cfg, monkeypatch):
+    monkeypatch.setattr(cfg, "PRINTERS", {"p1": "h"})
+    monkeypatch.setattr(cfg, "SERIALS", {})
+    monkeypatch.setattr(cfg, "LAN_KEYS", {"p1": "k"})
     with pytest.raises(RuntimeError):
-        bridge._validate_env()
+        cfg._validate_env()
